Fix grass shader link failure: match uMode/uDebug precision across stages
The grass program never linked on strict GL drivers (macOS ANGLE):
uMode and uDebug are declared in both shader stages, defaulting to
highp in the vertex shader and mediump in the fragment — a cross-stage
precision mismatch the driver rejects. Both now carry explicit mediump.
This was the entire invisible-grass bug; the one-time console
diagnostic surfaced it from the user's browser in one report.

+++ client/terrain-gl-layer.js
@@ -417,9 +417,9 @@ attribute vec3 aParams;   // phase, heightFrac (0 base, 1 tip), shade
 uniform vec3 uCamera;
 uniform vec2 uViewport;
 uniform float uTime;
-uniform float uMode;      // 0 = blade, 1 = ground shadow
-uniform vec3 uShadow;     // dirX, dirY, length
-uniform float uDebug;
+uniform mediump float uMode;      // 0 = blade, 1 = ground shadow
+uniform vec3 uShadow;             // dirX, dirY, length
+uniform mediump float uDebug;
 uniform float uWind;
 varying float vFrac;
 varying float vShade;
@@ -444,9 +444,9 @@ void main() {
 }`;
   const fragmentSource = `
 precision mediump float;
-uniform float uMode;
+uniform mediump float uMode;
 uniform float uDaylight;
-uniform float uDebug;
+uniform mediump float uDebug;
 varying float vFrac;
 varying float vShade;
 void main() {
